requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --index-url=https://pypi.org/simple
#
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.10
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2
    # via starlette
attrs==24.2.0
    # via aiohttp
click==8.1.7
    # via uvicorn
colorama==0.4.6
    # via click
fastapi==0.115.2
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   yarl
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
propcache==0.2.0
    # via yarl
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.23.4
    # via pydantic
pyyaml==6.0.2
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.39.2
    # via fastapi
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn==0.31.1
    # via -r requirements.in
yarl==1.15.2
    # via aiohttp
